gnome-shell: use nm-connection-editor
By default, the Network Settings uses some shitty control panel version which
sucks more ass than a porn star.  Let's change this back to the much better
nm-connection-editor.  Requires a patch tarball from me, or to be re-done so
that the patch is in FILESDIR instead of the patch tarball.

Signed-off-by: Steev Klimaszewski <steev@gentoo.org>